Updated supported models to include T7850
I have installed and tested Venstar T7850 (firmware VH 4.08-1A). It functions as it should on a heat-only 5-wire setup ... it has integrated nicely after activating the Local API feature in the thermostat menu. Just want to make the Home Assistant community aware that it is an option. I will also be testing it with an A/C unit in the next few weeks.